Split chat artifact panel boundaries
