+++ src/app/features/chat/artifacts/ArtifactPanel.tsx
@@ -1,28 +1,4 @@
-import {
-  BlockTypeSelect,
-  BoldItalicUnderlineToggles,
-  CodeToggle,
-  CreateLink,
-  DiffSourceToggleWrapper,
-  InsertCodeBlock,
-  InsertTable,
-  ListsToggle,
-  MDXEditor,
-  Separator,
-  UndoRedo,
-  codeBlockPlugin,
-  codeMirrorPlugin,
-  diffSourcePlugin,
-  headingsPlugin,
-  linkPlugin,
-  listsPlugin,
-  markdownShortcutPlugin,
-  quotePlugin,
-  tablePlugin,
-  thematicBreakPlugin,
-  toolbarPlugin,
-  type MDXEditorMethods,
-} from "@mdxeditor/editor";
+import { MDXEditor, type MDXEditorMethods } from "@mdxeditor/editor";
 import "@mdxeditor/editor/style.css";
 import {
   Download,
@@ -34,18 +10,12 @@ import {
   Play,
   Save,
 } from "lucide-react";
-import { useEffect, useMemo, useRef, useState } from "react";
-import ReactMarkdown from "react-markdown";
-import remarkGfm from "remark-gfm";
-import type { Artifact, ArtifactVersion } from "../../../desktop/types";
-import {
-  compileReactArtifactQuery,
-  listArtifactsQuery,
-  listArtifactVersionsQuery,
-  updateArtifactQuery,
-} from "../../../query/desktop-query";
+import { useEffect, useMemo, useRef } from "react";
 import { compactIconButtonClass } from "../../../ui/classes";
 import { cn } from "../../../utils/cn";
+import { createMarkdownEditorPlugins, HistoricalMarkdownPreview } from "./ArtifactMarkdown";
+import { formatArtifactSlug } from "./artifactFormat";
+import { useArtifactPanelState } from "./useArtifactPanelState";
 
 type ArtifactPanelProps = {
   conversationId: string | null;
@@ -55,148 +25,6 @@ type ArtifactPanelProps = {
   onClose: () => void;
 };
 
-type ArtifactView = "list" | "code" | "preview";
-
-function formatArtifactSlug(slug: string) {
-  return slug
-    .split("-")
-    .filter(Boolean)
-    .map((word) => `${word.slice(0, 1).toUpperCase()}${word.slice(1)}`)
-    .join(" ");
-}
-
-function escapeScriptContent(script: string) {
-  return script.replace(/<\/script/gi, "<\\/script");
-}
-
-function getArtifactExtension(kind: Artifact["kind"]) {
-  if (kind === "react") return "tsx";
-  if (kind === "markdown") return "md";
-  return "html";
-}
-
-const artifactScrollbarCss = `
-<style>
-  *::-webkit-scrollbar { width: 8px; height: 8px; }
-  *::-webkit-scrollbar-thumb { border: 1px solid transparent; border-radius: 999px; background: rgba(140, 148, 181, 0.22); background-clip: padding-box; }
-  * { scrollbar-color: rgba(140, 148, 181, 0.22) transparent; }
-</style>`;
-
-const artifactDarkPreviewCss = `
-    html { background: #262936; }
-    body { background: #262936; color: #d5daed; }
-    a { color: #b9bff3; }
-    code, pre { color: #d5daed; background: rgba(255,255,255,0.04); }
-    blockquote { color: #969db7; border-left: 3px solid rgba(185,191,243,0.32); margin-left: 0; padding-left: 1rem; }
-    hr { border: 0; border-top: 1px solid rgba(169,178,215,0.14); }
-`;
-
-function createMarkdownEditorPlugins(fullscreen: boolean, diffMarkdown: string) {
-  return [
-    headingsPlugin(),
-    listsPlugin(),
-    quotePlugin(),
-    linkPlugin(),
-    tablePlugin(),
-    thematicBreakPlugin(),
-    codeBlockPlugin({ defaultCodeBlockLanguage: "text" }),
-    codeMirrorPlugin({
-      codeBlockLanguages: {
-        css: "CSS",
-        html: "HTML",
-        js: "JavaScript",
-        jsx: "JavaScript JSX",
-        json: "JSON",
-        markdown: "Markdown",
-        md: "Markdown",
-        text: "Text",
-        ts: "TypeScript",
-        tsx: "TypeScript JSX",
-      },
-    }),
-    diffSourcePlugin({ viewMode: "rich-text", diffMarkdown }),
-    markdownShortcutPlugin(),
-    toolbarPlugin({
-      toolbarClassName: cn(
-        "artifact-mdx-toolbar",
-        fullscreen ? "artifact-mdx-toolbar-fullscreen" : "artifact-mdx-toolbar-drawer",
-      ),
-      toolbarContents: () => (
-        <DiffSourceToggleWrapper options={["rich-text", "source", "diff"]}>
-          <span className="artifact-mdx-toolbar-row artifact-mdx-toolbar-row-primary">
-            <UndoRedo />
-            <Separator />
-            <BlockTypeSelect />
-          </span>
-          <span className="artifact-mdx-toolbar-row artifact-mdx-toolbar-row-secondary">
-            <Separator />
-            <BoldItalicUnderlineToggles />
-            <CodeToggle />
-            <Separator />
-            <ListsToggle />
-            <CreateLink />
-            <InsertTable />
-            <InsertCodeBlock />
-          </span>
-        </DiffSourceToggleWrapper>
-      ),
-    }),
-  ];
-}
-
-function buildHtmlPreview(content: string) {
-  const capture = `${artifactScrollbarCss}
-<script>
-window.addEventListener('error', function(event) {
-  parent.postMessage({ source: 'howcode-artifact-preview', phase: 'runtime', message: event.message, stack: event.error && event.error.stack }, '*');
-});
-window.addEventListener('unhandledrejection', function(event) {
-  parent.postMessage({ source: 'howcode-artifact-preview', phase: 'runtime', message: String(event.reason && event.reason.message || event.reason), stack: event.reason && event.reason.stack }, '*');
-});
-const originalError = console.error;
-console.error = function(...args) {
-  parent.postMessage({ source: 'howcode-artifact-preview', phase: 'runtime', message: args.map(String).join(' ') }, '*');
-  originalError.apply(console, args);
-};
-</script>`;
-  return content.includes("<head")
-    ? content.replace(/<head([^>]*)>/i, `<head$1>${capture}`)
-    : `${capture}${content}`;
-}
-
-function buildReactPreview(compiledJs: string) {
-  return `<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  ${artifactScrollbarCss}
-  <style>
-    html, body, #root { min-height: 100%; margin: 0; }
-    body { font-family: Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif; }
-    ${artifactDarkPreviewCss}
-  </style>
-</head>
-<body>
-  <div id="root"></div>
-  <script>
-    window.addEventListener('error', function(event) {
-      parent.postMessage({ source: 'howcode-artifact-preview', phase: 'runtime', message: event.message, stack: event.error && event.error.stack }, '*');
-    });
-    window.addEventListener('unhandledrejection', function(event) {
-      parent.postMessage({ source: 'howcode-artifact-preview', phase: 'runtime', message: String(event.reason && event.reason.message || event.reason), stack: event.reason && event.reason.stack }, '*');
-    });
-    const originalError = console.error;
-    console.error = function(...args) {
-      parent.postMessage({ source: 'howcode-artifact-preview', phase: 'runtime', message: args.map(String).join(' ') }, '*');
-      originalError.apply(console, args);
-    };
-  </script>
-  <script type="module">${escapeScriptContent(compiledJs)}</script>
-</body>
-</html>`;
-}
-
 export function ArtifactPanel({
   conversationId,
   visible,
@@ -204,184 +32,40 @@ export function ArtifactPanel({
   onToggleFullscreen,
   onClose,
 }: ArtifactPanelProps) {
-  const [artifacts, setArtifacts] = useState<Artifact[]>([]);
-  const [selectedArtifactId, setSelectedArtifactId] = useState<string | null>(null);
-  const [view, setView] = useState<ArtifactView>("preview");
-  const [draft, setDraft] = useState("");
-  const [versions, setVersions] = useState<ArtifactVersion[]>([]);
-  const [selectedVersion, setSelectedVersion] = useState<number | "latest">("latest");
-  const [saving, setSaving] = useState(false);
-  const [previewHtml, setPreviewHtml] = useState("");
-  const [previewError, setPreviewError] = useState<string | null>(null);
-  const [previewRevision, setPreviewRevision] = useState(0);
-  const previousSelectedArtifactSlugRef = useRef<string | null>(null);
   const markdownEditorRef = useRef<MDXEditorMethods>(null);
-
-  const selectedArtifact = useMemo(
-    () =>
-      artifacts.find((artifact) => artifact.slug === selectedArtifactId) ?? artifacts[0] ?? null,
-    [artifacts, selectedArtifactId],
-  );
-  const selectedHistoricalVersion =
-    selectedVersion === "latest"
-      ? null
-      : (versions.find((version) => version.version === selectedVersion) ?? null);
-  const selectedArtifactSlug = selectedArtifact?.slug ?? null;
-  const selectedArtifactVersion = selectedArtifact?.version ?? null;
-  const displayedContent = selectedHistoricalVersion?.content ?? selectedArtifact?.content ?? "";
+  const panel = useArtifactPanelState(conversationId);
+  const {
+    artifacts,
+    selectedArtifact,
+    selectedVersion,
+    versions,
+    view,
+    draft,
+    displayedContent,
+    showingHistoricalVersion,
+    markdownPreviewEditable,
+    previewHtml,
+    previewError,
+    previewRevision,
+    saveDisabled,
+    saveDraft,
+    downloadArtifact,
+    setDraft,
+    setPreviewError,
+    setSelectedArtifactId,
+    setSelectedVersion,
+    setView,
+  } = panel;
   const markdownEditorPlugins = useMemo(
     () => createMarkdownEditorPlugins(fullscreen, selectedArtifact?.content ?? ""),
     [fullscreen, selectedArtifact?.content],
   );
-  const showingHistoricalVersion = Boolean(selectedHistoricalVersion);
-  const markdownPreviewEditable =
-    view === "preview" && selectedArtifact?.kind === "markdown" && !showingHistoricalVersion;
-  const saveDisabled =
-    !selectedArtifact ||
-    saving ||
-    view === "list" ||
-    (!showingHistoricalVersion && draft === selectedArtifact.content) ||
-    (view !== "code" && !markdownPreviewEditable && !showingHistoricalVersion);
-
-  useEffect(() => {
-    let cancelled = false;
-    setArtifacts([]);
-    setSelectedArtifactId(null);
-    setSelectedVersion("latest");
-    setVersions([]);
-    if (!conversationId) {
-      return;
-    }
-    void listArtifactsQuery(conversationId).then((nextArtifacts) => {
-      if (cancelled) return;
-      setArtifacts(nextArtifacts);
-      setSelectedArtifactId((current) =>
-        current && nextArtifacts.some((artifact) => artifact.slug === current)
-          ? current
-          : (nextArtifacts[0]?.slug ?? null),
-      );
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [conversationId]);
-
-  useEffect(() => {
-    if (!window.piDesktop?.subscribe) return;
-    return window.piDesktop.subscribe((event) => {
-      if (event.type !== "artifact-update") return;
-      if (!conversationId || event.conversationId !== conversationId) return;
-      setArtifacts((current) => {
-        const index = current.findIndex((artifact) => artifact.slug === event.artifact.slug);
-        if (index === -1) return [event.artifact, ...current];
-        const next = [...current];
-        next[index] = event.artifact;
-        return next;
-      });
-      setSelectedArtifactId(event.artifact.slug);
-      setSelectedVersion("latest");
-      setView("preview");
-      setPreviewRevision((revision) => revision + 1);
-    });
-  }, [conversationId]);
-
-  useEffect(() => {
-    if (previousSelectedArtifactSlugRef.current === selectedArtifactSlug) return;
-    previousSelectedArtifactSlugRef.current = selectedArtifactSlug;
-    setSelectedVersion("latest");
-  }, [selectedArtifactSlug]);
-
-  useEffect(() => {
-    let cancelled = false;
-    if (!selectedArtifactSlug) {
-      setVersions([]);
-      return;
-    }
-    void selectedArtifactVersion;
-    void listArtifactVersionsQuery(selectedArtifactSlug).then((nextVersions) => {
-      if (!cancelled) setVersions(nextVersions);
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [selectedArtifactSlug, selectedArtifactVersion]);
-
-  useEffect(() => {
-    setDraft(displayedContent);
-  }, [displayedContent]);
 
   useEffect(() => {
     if (view !== "preview" || selectedArtifact?.kind !== "markdown") return;
     if (markdownEditorRef.current?.getMarkdown() === displayedContent) return;
     markdownEditorRef.current?.setMarkdown(displayedContent);
   }, [displayedContent, selectedArtifact?.kind, view]);
-
-  useEffect(() => {
-    let cancelled = false;
-    setPreviewError(null);
-    if (!selectedArtifact) {
-      setPreviewHtml("");
-      return;
-    }
-    if (selectedArtifact.kind === "markdown") return;
-    if (selectedArtifact.kind === "html") {
-      setPreviewHtml(buildHtmlPreview(displayedContent));
-      return;
-    }
-    void compileReactArtifactQuery(displayedContent).then((result) => {
-      if (cancelled) return;
-      if (result.ok) {
-        setPreviewHtml(buildReactPreview(result.js));
-        setPreviewError(result.warnings.join("\n") || null);
-      } else {
-        setPreviewHtml("");
-        setPreviewError(result.error);
-      }
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [selectedArtifact, displayedContent]);
-
-  useEffect(() => {
-    const handleMessage = (event: MessageEvent) => {
-      if (event.data?.source !== "howcode-artifact-preview") return;
-      setPreviewError(
-        [event.data.phase, event.data.message, event.data.stack].filter(Boolean).join("\n"),
-      );
-    };
-    window.addEventListener("message", handleMessage);
-    return () => window.removeEventListener("message", handleMessage);
-  }, []);
-
-  const saveDraft = async () => {
-    if (!selectedArtifact) return;
-    const content = showingHistoricalVersion ? displayedContent : draft;
-    if (!showingHistoricalVersion && content === selectedArtifact.content) return;
-    setSaving(true);
-    try {
-      const updated = await updateArtifactQuery(selectedArtifact.slug, content, conversationId);
-      if (updated) {
-        setArtifacts((current) =>
-          current.map((artifact) => (artifact.slug === updated.slug ? updated : artifact)),
-        );
-        setSelectedVersion("latest");
-        setView("preview");
-        setPreviewRevision((revision) => revision + 1);
-      }
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const downloadArtifact = () => {
-    if (!selectedArtifact) return;
-    const content = showingHistoricalVersion ? displayedContent : draft;
-    void window.piDesktop?.saveTextToDownloads?.(
-      `${selectedArtifact.slug}.${getArtifactExtension(selectedArtifact.kind)}`,
-      content,
-    );
-  };
 
   if (!visible || !conversationId) return null;
 
@@ -542,12 +226,10 @@ export function ArtifactPanel({
           />
         ) : null}
 
-        {view === "preview" &&
-        selectedArtifact?.kind === "markdown" &&
-        !showingHistoricalVersion ? (
+        {markdownPreviewEditable ? (
           <div className="artifact-markdown-editor h-full min-h-0 overflow-hidden bg-[color:var(--sidebar)]">
             <MDXEditor
-              key={`${selectedArtifact.slug}:${selectedArtifact.version}`}
+              key={`${selectedArtifact?.slug}:${selectedArtifact?.version}`}
               ref={markdownEditorRef}
               markdown={draft}
               plugins={markdownEditorPlugins}
@@ -563,65 +245,7 @@ export function ArtifactPanel({
         ) : null}
 
         {view === "preview" && selectedArtifact?.kind === "markdown" && showingHistoricalVersion ? (
-          <div className="h-full min-h-0 overflow-auto bg-[color:var(--sidebar)] px-7 py-6 text-[14px] leading-[1.7] text-[color:var(--text)] [text-wrap:pretty] [&_h1]:[text-wrap:balance] [&_h2]:[text-wrap:balance] [&_h3]:[text-wrap:balance] [&_pre]:[text-wrap:initial]">
-            <ReactMarkdown
-              remarkPlugins={[remarkGfm]}
-              components={{
-                h1: ({ children }) => (
-                  <h1 className="mb-3 text-[20px] font-semibold text-[color:var(--text)]">
-                    {children}
-                  </h1>
-                ),
-                h2: ({ children }) => (
-                  <h2 className="mt-5 mb-2 text-[17px] font-semibold text-[color:var(--text)]">
-                    {children}
-                  </h2>
-                ),
-                h3: ({ children }) => (
-                  <h3 className="mt-4 mb-2 text-[15px] font-semibold text-[color:var(--text)]">
-                    {children}
-                  </h3>
-                ),
-                p: ({ children }) => <p className="my-2 text-[color:var(--text)]/92">{children}</p>,
-                ul: ({ children }) => <ul className="my-2 list-disc pl-5">{children}</ul>,
-                ol: ({ children }) => <ol className="my-2 list-decimal pl-5">{children}</ol>,
-                li: ({ children }) => (
-                  <li className="my-1 text-[color:var(--text)]/92">{children}</li>
-                ),
-                a: ({ children, href }) => (
-                  <a
-                    className="text-[color:var(--accent)] underline underline-offset-2"
-                    href={href}
-                    target={href?.startsWith("http") ? "_blank" : undefined}
-                    rel={href?.startsWith("http") ? "noreferrer" : undefined}
-                    onClick={(event) => {
-                      if (!href?.startsWith("http://") && !href?.startsWith("https://")) return;
-                      event.preventDefault();
-                      void window.piDesktop?.openExternal?.(href);
-                    }}
-                  >
-                    {children}
-                  </a>
-                ),
-                blockquote: ({ children }) => (
-                  <blockquote className="my-3 border-l-2 border-[rgba(185,191,243,0.32)] pl-4 text-[color:var(--muted)]">
-                    {children}
-                  </blockquote>
-                ),
-                code: ({ children }) => (
-                  <code className="font-mono text-[color:var(--accent)]">{children}</code>
-                ),
-                pre: ({ children }) => (
-                  <pre className="my-3 overflow-auto rounded-lg border border-[color:var(--border)] p-3 font-mono text-[12px] leading-5 text-[color:var(--text)]">
-                    {children}
-                  </pre>
-                ),
-                hr: () => <hr className="my-5 border-0 border-t border-[color:var(--border)]" />,
-              }}
-            >
-              {displayedContent}
-            </ReactMarkdown>
-          </div>
+          <HistoricalMarkdownPreview content={displayedContent} />
         ) : null}
 
         {view === "preview" && selectedArtifact?.kind !== "markdown" ? (
