+++ src/app/features/chat/artifacts/useArtifactPanelState.ts
@@ -1,0 +1,203 @@
+import { useEffect, useMemo, useRef, useState } from "react";
+import type { Artifact, ArtifactVersion } from "../../../desktop/types";
+import {
+  compileReactArtifactQuery,
+  listArtifactsQuery,
+  listArtifactVersionsQuery,
+  updateArtifactQuery,
+} from "../../../query/desktop-query";
+import { buildHtmlPreview, buildReactPreview } from "./artifactPreviewBuilders";
+import { getArtifactExtension } from "./artifactFormat";
+
+export type ArtifactView = "list" | "code" | "preview";
+
+export function useArtifactPanelState(conversationId: string | null) {
+  const [artifacts, setArtifacts] = useState<Artifact[]>([]);
+  const [selectedArtifactId, setSelectedArtifactId] = useState<string | null>(null);
+  const [view, setView] = useState<ArtifactView>("preview");
+  const [draft, setDraft] = useState("");
+  const [versions, setVersions] = useState<ArtifactVersion[]>([]);
+  const [selectedVersion, setSelectedVersion] = useState<number | "latest">("latest");
+  const [saving, setSaving] = useState(false);
+  const [previewHtml, setPreviewHtml] = useState("");
+  const [previewError, setPreviewError] = useState<string | null>(null);
+  const [previewRevision, setPreviewRevision] = useState(0);
+  const previousSelectedArtifactSlugRef = useRef<string | null>(null);
+
+  const selectedArtifact = useMemo(
+    () =>
+      artifacts.find((artifact) => artifact.slug === selectedArtifactId) ?? artifacts[0] ?? null,
+    [artifacts, selectedArtifactId],
+  );
+  const selectedHistoricalVersion =
+    selectedVersion === "latest"
+      ? null
+      : (versions.find((version) => version.version === selectedVersion) ?? null);
+  const selectedArtifactSlug = selectedArtifact?.slug ?? null;
+  const selectedArtifactVersion = selectedArtifact?.version ?? null;
+  const displayedContent = selectedHistoricalVersion?.content ?? selectedArtifact?.content ?? "";
+  const showingHistoricalVersion = Boolean(selectedHistoricalVersion);
+  const markdownPreviewEditable =
+    view === "preview" && selectedArtifact?.kind === "markdown" && !showingHistoricalVersion;
+  const saveDisabled =
+    !selectedArtifact ||
+    saving ||
+    view === "list" ||
+    (!showingHistoricalVersion && draft === selectedArtifact.content) ||
+    (view !== "code" && !markdownPreviewEditable && !showingHistoricalVersion);
+
+  useEffect(() => {
+    let cancelled = false;
+    setArtifacts([]);
+    setSelectedArtifactId(null);
+    setSelectedVersion("latest");
+    setVersions([]);
+    if (!conversationId) return;
+    void listArtifactsQuery(conversationId).then((nextArtifacts) => {
+      if (cancelled) return;
+      setArtifacts(nextArtifacts);
+      setSelectedArtifactId((current) =>
+        current && nextArtifacts.some((artifact) => artifact.slug === current)
+          ? current
+          : (nextArtifacts[0]?.slug ?? null),
+      );
+    });
+    return () => {
+      cancelled = true;
+    };
+  }, [conversationId]);
+
+  useEffect(() => {
+    if (!window.piDesktop?.subscribe) return;
+    return window.piDesktop.subscribe((event) => {
+      if (event.type !== "artifact-update") return;
+      if (!conversationId || event.conversationId !== conversationId) return;
+      setArtifacts((current) => {
+        const index = current.findIndex((artifact) => artifact.slug === event.artifact.slug);
+        if (index === -1) return [event.artifact, ...current];
+        const next = [...current];
+        next[index] = event.artifact;
+        return next;
+      });
+      setSelectedArtifactId(event.artifact.slug);
+      setSelectedVersion("latest");
+      setView("preview");
+      setPreviewRevision((revision) => revision + 1);
+    });
+  }, [conversationId]);
+
+  useEffect(() => {
+    if (previousSelectedArtifactSlugRef.current === selectedArtifactSlug) return;
+    previousSelectedArtifactSlugRef.current = selectedArtifactSlug;
+    setSelectedVersion("latest");
+  }, [selectedArtifactSlug]);
+
+  useEffect(() => {
+    let cancelled = false;
+    if (!selectedArtifactSlug) {
+      setVersions([]);
+      return;
+    }
+    void selectedArtifactVersion;
+    void listArtifactVersionsQuery(selectedArtifactSlug).then((nextVersions) => {
+      if (!cancelled) setVersions(nextVersions);
+    });
+    return () => {
+      cancelled = true;
+    };
+  }, [selectedArtifactSlug, selectedArtifactVersion]);
+
+  useEffect(() => {
+    setDraft(displayedContent);
+  }, [displayedContent]);
+
+  useEffect(() => {
+    let cancelled = false;
+    setPreviewError(null);
+    if (!selectedArtifact) {
+      setPreviewHtml("");
+      return;
+    }
+    if (selectedArtifact.kind === "markdown") return;
+    if (selectedArtifact.kind === "html") {
+      setPreviewHtml(buildHtmlPreview(displayedContent));
+      return;
+    }
+    void compileReactArtifactQuery(displayedContent).then((result) => {
+      if (cancelled) return;
+      if (result.ok) {
+        setPreviewHtml(buildReactPreview(result.js));
+        setPreviewError(result.warnings.join("\n") || null);
+      } else {
+        setPreviewHtml("");
+        setPreviewError(result.error);
+      }
+    });
+    return () => {
+      cancelled = true;
+    };
+  }, [selectedArtifact, displayedContent]);
+
+  useEffect(() => {
+    const handleMessage = (event: MessageEvent) => {
+      if (event.data?.source !== "howcode-artifact-preview") return;
+      setPreviewError(
+        [event.data.phase, event.data.message, event.data.stack].filter(Boolean).join("\n"),
+      );
+    };
+    window.addEventListener("message", handleMessage);
+    return () => window.removeEventListener("message", handleMessage);
+  }, []);
+
+  const saveDraft = async () => {
+    if (!selectedArtifact) return;
+    const content = showingHistoricalVersion ? displayedContent : draft;
+    if (!showingHistoricalVersion && content === selectedArtifact.content) return;
+    setSaving(true);
+    try {
+      const updated = await updateArtifactQuery(selectedArtifact.slug, content, conversationId);
+      if (updated) {
+        setArtifacts((current) =>
+          current.map((artifact) => (artifact.slug === updated.slug ? updated : artifact)),
+        );
+        setSelectedVersion("latest");
+        setView("preview");
+        setPreviewRevision((revision) => revision + 1);
+      }
+    } finally {
+      setSaving(false);
+    }
+  };
+
+  const downloadArtifact = () => {
+    if (!selectedArtifact) return;
+    const content = showingHistoricalVersion ? displayedContent : draft;
+    void window.piDesktop?.saveTextToDownloads?.(
+      `${selectedArtifact.slug}.${getArtifactExtension(selectedArtifact.kind)}`,
+      content,
+    );
+  };
+
+  return {
+    artifacts,
+    selectedArtifact,
+    selectedVersion,
+    versions,
+    view,
+    draft,
+    displayedContent,
+    showingHistoricalVersion,
+    markdownPreviewEditable,
+    previewHtml,
+    previewError,
+    previewRevision,
+    saveDisabled,
+    saveDraft,
+    downloadArtifact,
+    setDraft,
+    setPreviewError,
+    setSelectedArtifactId,
+    setSelectedVersion,
+    setView,
+  };
+}
